backend/pyproject.toml
[project]
name = "valora-backend"
version = "0.1.0"
description = "Backend service for the Valora system."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dateutil",
    "alembic",
    "fastapi",
    "resend",
    "google-auth",
    "PyJWT",
    "pydantic-settings",
    "psycopg[binary]",
    "requests",
    "simpleeval>=1.0.7,<2",
    "sqlalchemy",
    "uvicorn[standard]",
]

[dependency-groups]
dev = [
    "httpx",
    "pytest",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/valora_backend"]
